No backend ID in VM setup counts as unsuccessful
This changes the `create` subcommand where the VM instance being
created (using the startup script) has an issue where the
`tb_backend_id` label is created with no value (a failure state)
showed a 'success'. Now an empty `tb_backend_id` label will result in a
unsuccessful creation (and automatically deletes the VM instance that
was not properly setup.

PiperOrigin-RevId: 749831535